+++ app/domain/token_factory.rb
@@ -9,22 +9,23 @@ class TokenFactory < Dry::Struct
 
   attribute :slosilo, ::Types::Any.default{ Slosilo }
 
+  MAXIMUM_AUTHENTICATION_TOKEN_EXPIRATION = 5.hours
+
   def signing_key(account)
     slosilo["authn:#{account}".to_sym] || raise(NoSigningKey, account)
   end
 
-  def signed_host_key(account:, username:)
-    exp_t = Time.now + Rails.application.config.conjur_config.host_authorization_token_ttl.to_i
+  def signed_token(account:,
+                   username:,
+                   host_ttl: Rails.application.config.conjur_config.host_authorization_token_ttl,
+                   user_ttl: Rails.application.config.conjur_config.user_authorization_token_ttl)
+    offset = username.starts_with?('host/') ? host_ttl : user_ttl
+    exp_t = get_token_expiration(offset)
     signing_key(account).issue_jwt(sub: username, exp: exp_t)
   end
 
-  def signed_token(account:, username:)
-    if username.starts_with?('host/')
-      signed_host_key(account: account, username: username)
-    else
-      exp_t = Time.now + Rails.application.config.conjur_config.user_authorization_token_ttl.to_i
-      signing_key(account).issue_jwt(sub: username, exp: exp_t)
-    end
+  def get_token_expiration(offset)
+    offset < MAXIMUM_AUTHENTICATION_TOKEN_EXPIRATION ? Time.now + offset : Time.now + MAXIMUM_AUTHENTICATION_TOKEN_EXPIRATION
   end
 
 end